requirements.txt
Django==4.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.1
requests==2.31.0
